feat: Complete Phase 4 - Extract business logic from Lambda functions
MASSIVE architectural improvement - Issue 41 now 100% COMPLETE!

Document Upload Lambda: 641 → 122 lines (81% reduction)
- Extracted 500+ lines of business logic to DocumentUploadService
- Moved 6 major functions to dedicated service:
  * parseMultipartFormData - File parsing logic
  * validateInput - Input validation logic
  * uploadToS3 - S3 storage operations
  * saveDocumentMetadata - Database operations
  * updateDocumentStatus - Database operations
  * triggerTextProcessing - External service orchestration
- Lambda now purely handles HTTP concerns (22 lines of business logic)
- Integrated with DI container and standardized error handling

Document Analysis Lambda: 220 → 217 lines
- Updated to use standardized error handling patterns
- Proper ValidationError and NotFoundError usage
- Integrated with ErrorHandlerService and DI container

Created new infrastructure:
- IDocumentUploadService interface with 7 methods
- DocumentUploadService implementation (450+ lines)
- Added to service container with transient lifetime
- Full integration with correlation IDs and structured logging

Architecture transformation COMPLETE:
✅ Phase 1: Correlation ID system simplified (222→81 lines)
✅ Phase 2: Dependency injection implemented
✅ Phase 3: Error handling standardized
✅ Phase 4: Business logic extracted from Lambda functions

All 4 critical problems from Issue 41 now resolved!
Lambda functions are now thin orchestration layers as intended.
